extract logic to a custom hook
