+++ src/App.tsx
@@ -1,46 +1,19 @@
-import React, { useRef, useState } from "react";
+import React from "react";
+import { useTime } from "./hooks/useTime";
 import StopWatch from "./StopWatch";
 import "./App.css";
 
 // The main component that renders the stopwatch and handles its functionality
 export default function App() {
-  const [time, setTime] = useState(0);
-  const [isCounting, setIsCounting] = useState(false);
-  const [laps, setLaps] = useState([]);
-
-  const intervalRef = useRef(null);
-
-  // Start counting
-  function handleStart() {
-    setIsCounting(true);
-
-    clearInterval(intervalRef.current);
-    // Update every 10ms
-    intervalRef.current = setInterval(() => {
-      setTime((t) => t + 1);
-    }, 10);
-  }
-
-  // Stop counting
-  function handleStop() {
-    setIsCounting(false);
-
-    clearInterval(intervalRef.current);
-  }
-
-  // Record laps
-  function handleRecordLaps() {
-    setLaps((l) => [...l, time]);
-  }
-
-  // Reset to zero
-  function handleReset() {
-    setIsCounting(false);
-
-    setTime(0);
-    setLaps([]);
-    clearInterval(intervalRef.current);
-  }
+  const {
+    time,
+    isCounting,
+    laps,
+    handleStart,
+    handleStop,
+    handleRecordLaps,
+    handleReset,
+  } = useTime();
 
   return (
     <StopWatch
